Started on REST API with request handling
Request handling requires data to be able to be stored and accessed, so 'get' and 'put' requests can be used. I created a dictionary to be able to keep track of task IDs and store info on the name and date of tasks. I mainly worked on the back end of my project so far.